gamelit_litrpg_analysis/requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
schedule
pyyaml
vaderSentiment
nltk
streamlit
plotly